pyproject.toml
[tool.poetry]
name = "desmascarando-robos"
version = "0.1.0"
description = ""
authors = ["Gutto <guttolaudie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
ipython = "^8.24.0"
duckdb = "^0.10.2"
numpy = "^1.26.4"
pandas = "^2.2.2"
pyarrow = "^16.1.0"
scikit-learn = "^1.4.2"
plotly = "^5.22.0"
ipykernel = "^6.29.4"
nbformat = "^5.10.4"
setuptools = "^69.5.1"
ydata-profiling = "^4.8.3"
yellowbrick = "^1.5"
lightgbm = "^4.3.0"
xgboost = "^2.0.3"
catboost = "^1.2.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
black = "^24.4.2"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.4"
notebook = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
